% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_description}
\alias{check_description}
\title{Check if a package is identified as containing SIA modules}
\usage{
check_description(proj = curr_proj())
}
\arguments{
\item{proj}{\emph{character}, a path to the project. Defaults to \link[=curr_proj]{current project}.}
}
\value{
Side effects.
}
\description{
Check if a package is identified as containing SIA modules
}
\keyword{internal}
