% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{curr_proj}
\alias{curr_proj}
\title{Get a current project path}
\usage{
curr_proj()
}
\value{
The path to the current project.
}
\description{
This is a thin wrapper around \code{usethis::proj_get()} that silences any messages.
}
\examples{
\dontrun{
curr_proj()
}

}
\seealso{
Other helpers: 
\code{\link{default_shiny_io_functions}},
\code{\link{edit_rstudio_shortcuts}()},
\code{\link{list_categories}()},
\code{\link{open_sm_manifest}()},
\code{\link{print.sm_manifest}()},
\code{\link{sia_head_tag}()}
}
\concept{helpers}
