% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\docType{data}
\name{default_shiny_io_functions}
\alias{default_shiny_io_functions}
\title{Default \code{{shiny}} input/output UI functions consulted by
\code{module_namespace_linter()}}
\format{
An object of class \code{character} of length 25.
}
\usage{
default_shiny_io_functions
}
\description{
A character vector of function names whose calls are inspected for \code{ns()}
omission. Name of each element denotes the original package. Please refer to
\code{\link[=module_namespace_linter]{module_namespace_linter()}} for more details.
}
\details{
Following functions are covered:
\itemize{
\item \code{shiny::actionButton}
\item \code{shiny::actionLink}
\item \code{shiny::checkboxGroupInput}
\item \code{shiny::checkboxInput}
\item \code{shiny::dateInput}
\item \code{shiny::dateRangeInput}
\item \code{shiny::fileInput}
\item \code{shiny::numericInput}
\item \code{shiny::passwordInput}
\item \code{shiny::radioButtons}
\item \code{shiny::selectInput}
\item \code{shiny::sliderInput}
\item \code{shiny::textAreaInput}
\item \code{shiny::textInput}
\item \code{shiny::varSelectInput}
\item \code{shiny::dataTableOutput}
\item \code{shiny::tableOutput}
\item \code{shiny::uiOutput}
\item \code{shiny::htmlOutput}
\item \code{shiny::verbatimTextOutput}
\item \code{shiny::imageOutput}
\item \code{shiny::textOutput}
\item \code{shiny::plotOutput}
\item \code{plotly::plotlyOutput}
\item \code{DT::DTOutput}
}
}
\seealso{
Other helpers: 
\code{\link{curr_proj}()},
\code{\link{edit_rstudio_shortcuts}()},
\code{\link{list_categories}()},
\code{\link{open_sm_manifest}()},
\code{\link{print.sm_manifest}()},
\code{\link{sia_head_tag}()}
}
\concept{helpers}
\keyword{datasets}
