% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{edit_rstudio_shortcuts}
\alias{edit_rstudio_shortcuts}
\title{Show RStudio Keyboard Shortcuts}
\usage{
edit_rstudio_shortcuts()
}
\value{
No return value. Called for side effect.
}
\description{
Shows a popup window with RStudio keyboard shortcuts. Applicable only in
RStudio and in interactive \code{R} session.
}
\details{
You can quickly reach out solicited addin function by typing it in the
\code{Filter...} box in the very top of the window. Then double click at the blank
space just next to the addin function name and press down desired key or key
combination. Apply the changes and from now on, just call the addin function
with one keystroke. For more details, navigate to \href{https://support.posit.co/hc/en-us/articles/206382178-Customizing-Keyboard-Shortcuts-in-the-RStudio-IDE}{RStudio documentation}.
}
\examples{
if (interactive()) {
edit_rstudio_shortcuts()
}

}
\seealso{
Other helpers: 
\code{\link{curr_proj}()},
\code{\link{default_shiny_io_functions}},
\code{\link{list_categories}()},
\code{\link{open_sm_manifest}()},
\code{\link{print.sm_manifest}()},
\code{\link{sia_head_tag}()}
}
\concept{helpers}
