% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_categories.R
\name{list_categories}
\alias{list_categories}
\title{List the available SIA module categories}
\usage{
list_categories()
}
\value{
A character vector.
}
\description{
Lists all available categories a SIA module can be placed in. SIA app will
place the module with illegal category under "Modules".
}
\examples{
list_categories()

}
\seealso{
Other helpers: 
\code{\link{curr_proj}()},
\code{\link{default_shiny_io_functions}},
\code{\link{edit_rstudio_shortcuts}()},
\code{\link{open_sm_manifest}()},
\code{\link{print.sm_manifest}()},
\code{\link{sia_head_tag}()}
}
\concept{helpers}
