% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_sm_manifest}
\alias{open_sm_manifest}
\title{Open SIA Modules Manifest for Editing}
\usage{
open_sm_manifest(proj = curr_proj())
}
\arguments{
\item{proj}{\emph{character}, a path to the project. Defaults to \link[=curr_proj]{current project}.}
}
\value{
\emph{character}, a path to SIA Module Manifest (invisibly).
}
\description{
Open SIA Modules Manifest for Editing
}
\examples{
\dontrun{
open_sm_manifest()
}

}
\seealso{
Other helpers: 
\code{\link{curr_proj}()},
\code{\link{default_shiny_io_functions}},
\code{\link{edit_rstudio_shortcuts}()},
\code{\link{list_categories}()},
\code{\link{print.sm_manifest}()},
\code{\link{sia_head_tag}()}
}
\concept{helpers}
