% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modules.R
\name{print.sm_manifest}
\alias{print.sm_manifest}
\title{Print a SIA Modules Manifest}
\usage{
\method{print}{sm_manifest}(x, ..., as_tibble = FALSE)
}
\arguments{
\item{x}{\emph{sm_manifest} object, i.e., an output of \code{\link[=get_modules]{get_modules()}}.}

\item{...}{Not used at the moment.}

\item{as_tibble}{\emph{logical}, print the manifest as a \code{tibble}? Defaults to
\code{FALSE}.}
}
\value{
Called for side effects by default. Returns a \code{tibble} if \code{as_tibble}
argument is set to \code{TRUE}.
}
\description{
Prints out the SIA Modules Manifest obtained through \code{\link[=get_modules]{get_modules()}}. By
default, a plain YAML content is returned, but you can also get a formatted
output in a \code{tibble}, which is suitable for packages with large number of SIA
modules.
}
\seealso{
Other helpers: 
\code{\link{curr_proj}()},
\code{\link{default_shiny_io_functions}},
\code{\link{edit_rstudio_shortcuts}()},
\code{\link{list_categories}()},
\code{\link{open_sm_manifest}()},
\code{\link{sia_head_tag}()}
}
\concept{helpers}
