% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_module.R
\name{remove_module}
\alias{remove_module}
\title{Remove a module}
\usage{
remove_module(module_id = NULL, proj = curr_proj())
}
\arguments{
\item{module_id}{\emph{character}, name of the module to remove (including the
prefix). If \code{NULL} (the default), all modules discovered by \code{\link[=get_modules]{get_modules()}}
are listed and you are asked to pick one.}

\item{proj}{\emph{character}, a path to the project. Defaults to \link[=curr_proj]{current project}.}
}
\value{
No return value. Called for the side effect.
}
\description{
Removes the given module from the SIA Module Manifest and deletes the
respective \code{.R} file.
}
\examples{
\dontrun{
remove_module()
}


}
\seealso{
Other module management functions:
\code{\link{add_module}()},
\code{\link{get_modules}()},
\code{\link{preview_module}()}
}
\concept{module_management}
