% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_module.R
\name{save_and_document_fun}
\alias{save_and_document_fun}
\title{Save all opened files and document the package in development}
\usage{
save_and_document_fun(proj = curr_proj())
}
\arguments{
\item{proj}{\emph{character}, a path to the project. Defaults to \link[=curr_proj]{current project}.}
}
\value{
No return value. Called for side effects.
}
\description{
Save all opened files and document the package in development
}
\keyword{internal}
