% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SteelDwassTest.R
\name{SDtest}
\alias{SDtest}
\title{Performing Steel-Dwass test}
\usage{
SDtest(dt, save_results)
}
\arguments{
\item{dt}{data set with ID on column 1 and stable isotope ratios after on column 2}

\item{save_results}{if you want to save results as .txt file, put TRUE here}
}
\description{
\code{SDtest} set the slack information to system
}
\examples{
# mydata <- read.csv("All.csv", header = T)
# SDtest(mydata, TRUE)
}
