% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_module.R
\name{sia_head_tag}
\alias{sia_head_tag}
\title{Core HTML head tag from ShinyItemAnalysis}
\usage{
sia_head_tag()
}
\value{
HTML head tag of class \code{shiny.tag}.
}
\description{
This function is intended to be used for SIA module preview only, which is
done by default in \code{\link[=preview_module]{preview_module()}}. It provides a HTML head tag with a
math rendering facility provided by the
\ifelse{html}{\eqn{\KaTeX}}{\emph{KaTeX}} library, and some CSS rules to
format tables. \emph{If printed in the console, expect nothing to be shown.}
}
\seealso{
Other helpers: 
\code{\link{curr_proj}()},
\code{\link{default_shiny_io_functions}},
\code{\link{edit_rstudio_shortcuts}()},
\code{\link{list_categories}()},
\code{\link{open_sm_manifest}()},
\code{\link{print.sm_manifest}()}
}
\concept{helpers}
