% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIAtools-package.R
\docType{package}
\name{SIAtools-package}
\alias{SIAtools}
\alias{SIAtools-package}
\title{SIAtools: 'ShinyItemAnalysis' Modules Development Toolkit}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A comprehensive suite of functions designed for constructing and managing 'ShinyItemAnalysis' modules, supplemented with detailed guides, ready-to-use templates, linters, and tests. This package allows developers to seamlessly create and integrate one or more modules into their existing packages or to start a new module project from scratch.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://applstat.github.io/SIAtools/}
}

}
\author{
\strong{Maintainer}: Jan Netik \email{netik@cs.cas.cz} (\href{https://orcid.org/0000-0002-3888-3203}{ORCID})

Authors:
\itemize{
  \item Patricia Martinkova \email{martinkova@cs.cas.cz} (\href{https://orcid.org/0000-0003-4754-8543}{ORCID})
}

}
\keyword{internal}
