% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sm_list_available}
\alias{sm_list_available}
\title{List SIA module packages available on the official repository}
\usage{
sm_list_available(repo = default_repo())
}
\arguments{
\item{repo}{\emph{character}, a URL to the repository. Defaults to the official
SIA repository.}
}
\description{
List SIA module packages available on the official repository
}
\keyword{internal}
