% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WilcoxonTest.R
\name{WCtest}
\alias{WCtest}
\title{Performing Wilcoxon test}
\usage{
WCtest(dt, save_results)
}
\arguments{
\item{dt}{data set with ID (column 1) and stable isotope ratios (column 2~n)}

\item{save_results}{if you want to save results as .txt file, put TRUE here}
}
\description{
\code{WCtest} set the slack information to system
}
\examples{
# mydata <- read.csv("All.csv", header = T)
# WCtest(mydata, TRUE)
}
