% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_module_project.R
\name{welcome_routine}
\alias{welcome_routine}
\title{Do a welcome routine in the fresh SIA module project}
\usage{
welcome_routine()
}
\description{
Do a welcome routine in the fresh SIA module project
}
\keyword{internal}
